+++ Cecilifier.Web/CecilifierRequest.cs
@@ -1,0 +1,10 @@
+using System.Text.Json.Serialization;
+
+namespace Cecilifier.Web
+{
+    public class CecilifierRequest
+    {
+        [JsonPropertyName("options")] public WebOptions WebOptions { get; set; }
+        [JsonPropertyName("code")] public string Code { get; set; }
+    }
+}