+++ CHANGELOG.md
@@ -1,6 +1,27 @@
 ## unreleased 
 
-## 13/July/2021
+## Changed
+
+## Added
+
+## 04/July/2021
+## Changed
+
+- bump cecilifier version to 1.13.0
+- fixed a couple of ref parameter issues
+- updated Microsoft.CodeAnalysis.CSharp to version 3.10.0-3.final
+- updated Mono.Cecil to version 0.11.3
+- updated Newtonsoft.Json to version 13.0.1
+- changed release notes UI.
+
+## Added
+
+- support for *out variables*
+- support for *ref return*
+- support for *local ref*
+- code coverage
+
+## 13/June/2021
 
 ## Changed
 
@@ -33,7 +54,11 @@
 ## Added
 
 - a lot of other stuff that has not been reported since March/2019 :(
+<<<<<<< HEAD
 - Support for for statements
+=======
+- Support for for statements	
+>>>>>>> 70ebeb1 (updates Cecil & Microsoft.CodeAnalysis.CSharp versions)
 
 ## 24/March/2019
 
@@ -56,4 +81,8 @@
 
 ### Added
 
+<<<<<<< HEAD
 - Support access of generic static/instance methods.
+=======
+- Support access of generic static/instance methods.
+>>>>>>> 70ebeb1 (updates Cecil & Microsoft.CodeAnalysis.CSharp versions)